host: Fetch the client every time.
It's just a pointer, so we shouldn't really worry. With the old code path, I
think some cases never would have been reached in the defer, we got a client if
there was Stdin or if we had a TTY, but then the defer checks for a client and a
TTY or a Stream, so I believe that in the case of a stream without stdin, we'd
skip that path.

Signed-off-by: Blaž Hrastnik <blaz.hrast@gmail.com>